requirements.txt
streamlit>=1.20
pandas>=1.4
altair>=4.2
python-dateutil>=2.8